requirements.txt
numpy==1.25.2
opencv-python==4.4.0.46
Pillow==10.3.0
streamlit_cookies_controller==0.0.4
